add new window to create a db file

+++ controllers/home_controller.py
@@ -4,17 +4,18 @@ from tkinter.ttk import Entry, Button
 import ttkbootstrap as ttk
 from pykeepass import PyKeePass
 
+from controllers.new_db_file_controller import CreateFile
 
 class OpenFile():
     def __init__(self):
-        self.root = ttk.Window(themename='darkly')
+        self.root = ttk.Window(themename='cyborg')
         self.geometry = self.root.geometry("650x300")
 
         # FILE PATH FRAME
         self.filepath_frame = ttk.Frame(self.root)
         self.filepath_label = ttk.Label(self.filepath_frame, text="Select a file", width=10)
-        self.filepath_entry = Entry(self.filepath_frame,bootstyle="info", width=40)
-        self.filepath_button = Button(self.filepath_frame,bootstyle="info-outline", text='Select',
+        self.filepath_entry = Entry(self.filepath_frame, style="info", width=40)
+        self.filepath_button = Button(self.filepath_frame, style="info-outline", text='Select',
                                       command=self.filepath_button, width=9)
         self.filepath_label.grid(row=0, column=0, padx=20, sticky="W")
         self.filepath_entry.grid(row=0, column=1, padx=20, sticky="W")
@@ -24,8 +25,8 @@ class OpenFile():
         # PASSWORD FRAME
         self.password_frame = ttk.Frame(self.root)
         self.password_label = ttk.Label(self.password_frame, text="Password",width=10)
-        self.password_entry = Entry(self.password_frame, bootstyle='danger', show='*', width=40)
-        self.password_button = Button(self.password_frame, bootstyle="danger-outline", text="Ok",
+        self.password_entry = Entry(self.password_frame, style='danger', show='*', width=40)
+        self.password_button = Button(self.password_frame, style="danger-outline", text="Ok",
                                       command=self.password_button, width=9)
         self.password_label.grid(row=0, column=0, padx=20, sticky="W")
         self.password_entry.grid(row=0, column=1, padx=20, sticky="W")
@@ -43,23 +44,24 @@ class OpenFile():
 
         # NEW FILE FRAME
         self.new_file_frame = ttk.Frame(self.root)
-        self.new_file_button = Button(self.new_file_frame, bootstyle='succes',
-                                      text="Create a new database for your passwords", command=None, width=73)
+        self.new_file_button = Button(self.new_file_frame, style='succes',
+                                      text="Create a new database for your passwords", command=self.create_new_db, width=73)
         self.new_file_button.grid(row=0, column=0, padx=20, sticky="W")
         self.new_file_frame.grid(row=0, column=0, padx=20)
         self.new_file_frame.place(y=235)
 
     def filepath_button(self):
+        """Get file path & check .kdbx format"""
         path = filedialog.askopenfilename()
+        if not Path(path).suffix == ".kdbx":
+            return self.wrong_password_label.config(text='Please select a .kdbx file')
+
         self.filepath_entry.delete(0, "end")
         self.filepath_entry.insert(0, path)
         return path
 
-    def read_entries(self, entries):
-        for k,v in entries:
-            print(f"{k.upper} : {v}")
-
     def password_button(self):
+        """Check Password & if OK return entries"""
         password = self.password_entry.get()
         if password == "":
             return self.wrong_password_label.config(text='Enter a password')
@@ -67,9 +69,6 @@ class OpenFile():
         db_path = self.filepath_entry.get()
         if not db_path or not Path(db_path).exists():
             return self.wrong_password_label.config(text='Select a file first')
-
-        if not Path(db_path).suffix == ".kdbx":
-            return self.wrong_password_label.config(text='Please select a .kdbx file')
 
         try:
             kp = PyKeePass(self.filepath_entry.get(), password=password)
@@ -84,6 +83,10 @@ class OpenFile():
             self.wrong_password_label.config(text='Wrong Password try again')
             raise e
 
+    def create_new_db(self):
+        """Open new window to create_db form"""
+        create_db = CreateFile(self.root)
+        create_db.run()
+
     def run(self):
         self.root.mainloop()
-
